[0.0.1+3] Improved controller handling, updated veto package, started with MapInformer example
