+++ example/lib/widgets/examples/method_example.dart
