+++ example/lib/widgets/util/shrink.dart
